requirements.txt
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.1
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.65.5
idna==3.7
joblib==1.4.2
lxml==5.3.0
nltk==3.8.1
pdf2image==1.17.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-docx==1.1.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
